requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-google-genai
pypdf
tiktoken
